feat: remove WebSocket-based market making example; enhance spread strategy to support dynamic market making with improved logging and order management
